+++ modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/exec/ddl/DdlCommandHandlerExceptionHandlingTest.java
@@ -25,8 +25,8 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import java.util.concurrent.CompletableFuture;
 import org.apache.ignite.internal.catalog.CatalogManager;
 import org.apache.ignite.internal.catalog.CatalogTestUtils;
-import org.apache.ignite.internal.distributionzones.DistributionZoneAlreadyExistsException;
-import org.apache.ignite.internal.distributionzones.DistributionZoneNotFoundException;
+import org.apache.ignite.internal.catalog.DistributionZoneExistsValidationException;
+import org.apache.ignite.internal.catalog.DistributionZoneNotFoundValidationException;
 import org.apache.ignite.internal.hlc.HybridClockImpl;
 import org.apache.ignite.internal.sql.engine.prepare.ddl.CreateZoneCommand;
 import org.apache.ignite.internal.sql.engine.prepare.ddl.DropZoneCommand;
@@ -60,7 +60,7 @@ public class DdlCommandHandlerExceptionHandlingTest extends IgniteAbstractTest {
 
     @Test
     public void testZoneAlreadyExistsOnCreate1() {
-        assertThat(handleCreateZoneCommand(false), willThrow(DistributionZoneAlreadyExistsException.class));
+        assertThat(handleCreateZoneCommand(false), willThrow(DistributionZoneExistsValidationException.class));
     }
 
     @Test
@@ -73,7 +73,7 @@ public class DdlCommandHandlerExceptionHandlingTest extends IgniteAbstractTest {
         DropZoneCommand cmd = new DropZoneCommand();
         cmd.zoneName(ZONE_NAME);
 
-        assertThat(commandHandler.handle(cmd), willThrow(DistributionZoneNotFoundException.class));
+        assertThat(commandHandler.handle(cmd), willThrow(DistributionZoneNotFoundValidationException.class));
     }
 
     @Test
